requirements.txt
simple-salesforce==1.12.6
python-dotenv==1.0.0
jinja2==3.1.2
pandas==2.1.4
requests==2.31.0
